Moved CustomUser and removed spring's user extension
